+++ src/main/kotlin/budget/BudgetSummary.kt
@@ -5,8 +5,8 @@ data class BudgetSummary(val id: Int,
                          val reference: String, // reference comptable
                          val type: String, // e.g. fonctionnement, investissement
                          val recipient: String, // e.g. maternelle, primaire, général
-                         val realRemaining: Float = 0.0f, // reste réel (commandes en cours non prise en compte)
-                         val virtualRemaining: Float = 0.0f // reste virtuel (commandes en cours déduites)
+                         val realRemaining: Double = 0.0, // reste réel (commandes en cours non prise en compte)
+                         val virtualRemaining: Double = 0.0 // reste virtuel (commandes en cours déduites)
 ) {
     companion object {
         fun createFromBudget(budget: Budget): BudgetSummary {
